+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - Improve fastlane lane test_all
+- Add examples for watchOS and tvOS
 
 ## [0.1.0] - 2023-11-11